+++ config/theme-setup.php
@@ -15,28 +15,38 @@ return array(
 	// Theme support features to add
 	//=============================================
 	'add_theme_support' => array(
-		'html5'                       => array(
+		'html5'                           => array(
 			'caption',
 			'comment-form',
 			'comment-list',
+			'gallery',
 			'search-form'
 		),
-		'genesis-responsive-viewport' => null,
-		'genesis-footer-widgets'      => 4,
-		'genesis-structural-wraps'    => array(
-			'footer',
-			'footer-widgets',
-			'header',
-			'nav',
-			'site-inner',
-			'site-tagline',
-			'partners__footer',
+		'genesis-accessibility'           => array(
+			'404-page',
+			'drop-down-menu',
+			'headings',
+			'rems',
+			'search-form',
+			'skip-links'
 		),
-		'genesis-menus'               => array(
-			'primary' => __( 'Primary Navigation Menu', CHILD_TEXT_DOMAIN ),
-			'secondary' => __( 'Secondary Navigation Menu', CHILD_TEXT_DOMAIN ),
-			'footer'  => __( 'Footer Navigation Menu', CHILD_TEXT_DOMAIN ),
+		'genesis-responsive-viewport'     => null,
+		'custom-header'                   => array(
+			'width'           => 600,
+			'height'          => 160,
+			'header-selector' => '.site-title a',
+			'header-text'     => false,
+			'flex-height'     => true,
 		),
+		'custom-background'               => null,
+		'genesis-after-entry-widget-area' => null,
+		'genesis-footer-widgets'          => 3,
+		'genesis-menus'                   => array(
+			'primary'   => __( 'Primary Navigation', CHILD_TEXT_DOMAIN ),
+			'secondary' => __( 'Secondary Navigation', CHILD_TEXT_DOMAIN ),
+			'tertiary' => __( 'Footer Navigation', CHILD_TEXT_DOMAIN )
+		),
+		'woocommerce',
 //		'genesis-after-entry-widget-area' => null,
 	),
 	//=============================================
@@ -53,11 +63,21 @@ return array(
 	// Theme Default Settings
 	//=============================================
 	'theme_default_settings' => array(
-		'blog_cat_num'              => 10,
+		'blog_cat_num'              => 12,
 		'content_archive'           => 'full',
-		'content_archive_limit'     => 250,
+		'content_archive_limit'     => 0,
 		'content_archive_thumbnail' => 0,
 		'posts_nav'                 => 'numeric',
 		'site_layout'               => 'full-width-content',
+	),
+	//=============================================
+	// Image Sizes
+	//=============================================
+	'theme_image_sizes' => array(
+		'featured-image' => array(
+			'width'  => 720,
+			'height' => 400,
+			'crop'   => true,
+		),
 	),
 );